[WIP] Make npdependency an installable package (#2)
# Changes

## Added

- setuptools machinery to make this an installable package ([b0a68b5](https://github.com/bencrabbe/npdependency/commit/b0a68b5152895304dbc38bf7a0276656235a11cd))
- console entry point for `graph_parser3.py` which is now invokable by the command `graph_parser` when the npdependnecy package is installed ([1ed4e4c](https://github.com/bencrabbe/npdependency/commit/1ed4e4cf9c3760c82a9a60ba6635335c95353924))
- a version number (0.1.0), I suggest using [semver](https://semver.org/) for future versions of the package

## Changed

- The `src` dir has been renamed to `npdependency` because setuptools [makes](https://github.com/pypa/setuptools/issues/250) everything harder when a package's parent directory is not named after the package
- The readme has been changed to reflect the changes here and now advises to install this as a package (for now using pip and the git http url, but we should probably release it on Pypi at some point) and use the console entry point instead of calling the script directly (although that is still possible, although the path using `src` have to be rewritten, see above)
- `DependencyDataset` is now in `deptree` rather than directly in `graph_parser3`

## Removed

- The Profiterole and SPMRL stuff, which is not actually related (I think?) to the parser. It still lives in the git history anyway, but we could make it a dedicated repo (in fact, I already have a repo for the MCVF conversion stuff somewhere with improvements)
