+++ requirements.txt
@@ -1,4 +1,4 @@
 fasttext
 torch
 transformers
-yaml
+pyyaml
