+++ setup.cfg
@@ -1,0 +1,26 @@
+[metadata]
+name = npdependency
+version = 0.1.0
+description = A graph parser to tinker with
+license = MIT
+author = Benoît Crabbé
+author_email = benoit.crabbe@linguist.univ-paris-diderot.fr
+long_description = file: README.md
+long_description_content_type = text/markdown
+keywords =
+    nlp
+    parsing
+
+[options]
+packages = find:
+include_package_data=True
+python_requires = >= 3.7
+install_requires =
+    fasttext
+    torch >= 1.6, < 2.0.0
+    transformers >= 3.2.0, < 4.0.0
+    pyyaml
+
+[options.entry_points]
+console_scripts =
+    graph_parser = npdependency.graph_parser3:main
