--- spmrl/README
@@ -1,1 +1,0 @@
-That's where you put the treebank files
